renamed files, cleanup some variables
