+++ src/event-scripts/validate-config.ts
